third_party/blink/web_tests/platform/generic/fast/forms/select/select-clientheight-with-multiple-attr-expected.txt

HTMLSelectElement multiple attribute test when no size is specified

PASS clientHeight('sel2') is clientHeight('sel1')
PASS getElemById('sel2').setAttribute('size', '5'); clientHeight('sel2') > clientHeight('sel1') is true
PASS clientHeight('sel3') is clientHeight('sel1')
PASS getElemById('sel3').setAttribute('size', '2'); clientHeight('sel3') < clientHeight('sel1') is true
PASS getElemById('sel3').setAttribute('size', '0'); clientHeight('sel3') is clientHeight('sel1')
PASS getElemById('sel3').setAttribute('size', ''); clientHeight('sel3') is clientHeight('sel1')
PASS getElemById('sel3').setAttribute('size', '1+ef'); clientHeight('sel3') < clientHeight('sel1') is true
PASS successfullyParsed is true

TEST COMPLETE

One
 
One
Two
Three
Four
Five
Six
Seven
Eight
Nine
Ten
Eleven
Twelve
Thirteen
Fourteen
 
One
Two
Three
